Add voice recognition service and UI components
- Implemented VoiceRecognitionService for processing speech recognition data, keyword detection, and delayed saving of keywords.
- Created VoiceRecognitionUIService to manage UI state and interactions with the voice recognition service.
- Developed voice recognition widgets for displaying recognized texts, recording controls, keyword indicators, and class selection.
- Integrated Google Calendar event registration based on recognized keywords and detected dates.
